Add difference solution

+++ CODEFORCES SHEET 1/C.SimpleCalculator.cpp
@@ -14,7 +14,7 @@
 using namespace std;
 
 int main(){
-    int X, Y;
+    long long X, Y;
     cin >> X >> Y;
 
     cout << X << " + " << Y << " = " << X + Y << endl;
